Do not request mouse events from the terminal for now
Currently, the mouse isn't supported, but mouse events are requested
from the terminal. This causes some unexpected behavior. This commit
stops requesting mouse events.

Once mouse support is implemented (which would be great!), there should
probably be an option to disable it for the same reason.

This will help with https://github.com/blacknon/hwatch/issues/51 and
https://github.com/blacknon/hwatch/issues/66.