Restore terminal if `hwatch` is killed
Currently, if `hwatch` is running in one terminal and 
you run `killall hwatch` from another, the first terminal
ends up with an invisible cursor. A workaround I found
is to run `tput reset` in that terminal.

This commit uses the [Ctrl-C crate] to make `hwatch`
restore the terminal to a sane state even if it's killed
or otherwise interrupted (within limits, `killall -9 hwatch`
will not let `hwatch` do anything before being killed).

[Ctrl-C crate]: https://crates.io/crates/ctrlc