Watch diff: fix display: change space from delete to nbsp
Changes the space character used for padding and as a last
character from the delete character (U+007F) to nbsp (U+00A0).

The delete character causes some issues when appended as the last
character by in the Watch diff mode (the last character from the
date is seen), changing the space character to a nbsp fixes this.

Fixes blacknon/hwatch#88.